Zombies spawn during the day, moved files to com.mortuusterra, modified zombie burn algo
