+++ src/com/mortuusterra/utils/files/FileType.java
@@ -1,0 +1,7 @@
+package com.mortuusterra.utils.files;
+
+public enum FileType {
+    YAML,
+    JSON,
+    TEXT
+}
