--- src/mortuusterra/objects/tower/CellTowerObject.java
@@ -1,18 +1,0 @@
-package mortuusterra.objects.tower;
-
-import org.bukkit.Location;
-
-public class CellTowerObject {
-	
-	private Location towerLocation;
-	
-	
-	public CellTowerObject(Location towerLocation) {
-		this.towerLocation = towerLocation;
-	}
-	
-	public Location getTowerLocation() {
-		return this.towerLocation;
-	}
-
-}
