--- src/mortuusterra/utils/files/FileType.java
@@ -1,7 +1,0 @@
-package mortuusterra.utils.files;
-
-public enum FileType {
-    YAML,
-    JSON,
-    TEXT
-}
